+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Improvements
 - PR #103 DGA detector refactor
+- PR #120 Use pytest tmpdir fixtures in unit tests
 
 ## Bug Fixes